refactor(terminal): 拆出时间戳环/链接下划线/命令块 gutter hook 与共享类型
- 新增 components/terminal/：terminalTypes（TerminalProps/CommandBlockState/全局快照声明）、useTerminalTimestamps（marker ring 时间戳管理）、useTerminalLinkUnderlines（可见区链接扫描缓存与常驻下划线层）、useTerminalGutter（命令块收起/展开与时间戳侧栏渲染）
- Terminal.tsx 4233 -> 3568 行
