+++ frontend/src/components/Terminal.tsx
@@ -24,6 +24,10 @@ import {
 } from '../utils/terminalCommandAutocomplete.ts';
 import { parseCommandInputContext } from '../utils/terminalCommandAutocompleteParser.ts';
 import Tiptop from './Tiptop.tsx';
+import { useTerminalTimestamps } from './terminal/useTerminalTimestamps.ts';
+import { useTerminalLinkUnderlines } from './terminal/useTerminalLinkUnderlines.ts';
+import { useTerminalGutter } from './terminal/useTerminalGutter.ts';
+import type { TerminalProps } from './terminal/terminalTypes.ts';
 import { ToggleSwitch } from './settings/SharedComponents.tsx';
 import type { QuickCommandsHandle } from './QuickCommands.tsx';
 import { Button, ContextMenu, Modal } from './ui';
@@ -36,41 +40,11 @@ import { getTerminalTheme, getAppThemeMode, isDarkTerminalSurface, getSolidTermi
 import { getResolvedProgramFontPreferences } from '../utils/programFonts.ts';
 import { highlightKeywords, loadKeywordRulesFromStorage, setKeywordRules, createHighlightState, type KeywordRule } from '../utils/terminalKeywordHighlight.ts';
 
-// 终端快照供 FileManager 等模块读取当前 buffer 文本（FileManager.jsx 亦写入同名键）
-declare global {
-  interface Window {
-    __luminTerminalSnapshots?: Record<string, () => string>;
-  }
-}
-
-/** 时间戳 ring 条目 / 命令块状态 */
-interface CommandBlockState {
-  id: number;
-  commandLineText: string;
-  occurrence: number;
-  collapsed: boolean;
-  savedOutput: string[] | null;
-  savedOutputTs: string[] | null;
-}
 
 // 启动时从 localStorage 加载自定义关键字规则（模块级，仅执行一次）
 loadKeywordRulesFromStorage();
 
-export interface TerminalProps {
-  sessionId: string;
-  serverId: string;
-  historyServerId: string;
-  status: string;
-  isActive: boolean;
-  serverName: string;
-  connectedSessions?: Array<{ id?: string }>;
-  showCommands?: boolean;
-  onQuickCommandsOpenChange?: (open: boolean) => void;
-  quickCmdsRef?: React.RefObject<QuickCommandsHandle | null>;
-  // 重连触发器：串口/本地复用同一 sessionId 重连时，wsRebuildKey 自增，
-  // 让下方建立 xterm+WebSocket 的主 effect 重跑，重建 WS（对齐 SSH 重连靠新 terminalId 触发的行为）。
-  wsRebuildKey?: number;
-}
+export type { TerminalProps } from './terminal/terminalTypes.ts';
 
 export default function Terminal({
   sessionId,
@@ -264,669 +238,29 @@ export default function Terminal({
   const alternateBufferActiveRef = useRef(false);
   const screenScrollbackRef = useRef({ pending: false, active: false });
   const prepareScreenScrollbackRef = useRef<(command: string) => void>(() => {});
-  // Ring buffer 时间戳：用 xterm marker 跟随 scrollback 裁剪，避免 buffer 行号复用后错位
-  const TS_POOL = 6000;
-  // null! 惰性初始化惯用法：下方 if 守卫保证首次渲染即完成填充，后续恒非空
-  const tsRingRef = useRef<{ entries: Array<{ marker: IMarker; val: string } | null>; next: number }>(null!);
-  if (!tsRingRef.current) {
-    tsRingRef.current = { entries: new Array(TS_POOL), next: 0 };
-  }
-  const tsSet = (marker: IMarker | undefined, val: string) => {
-    if (!marker) return;
-    const r = tsRingRef.current;
-    // 同 line 只保留最新戳（执行命令时要盖掉空提示符上的旧时间）
-    const line = marker.line;
-    if (typeof line === 'number' && line >= 0) {
-      for (let j = 0; j < r.entries.length; j += 1) {
-        const entry = r.entries[j];
-        if (entry?.marker?.line === line) {
-          entry.marker.dispose?.();
-          r.entries[j] = null;
-        }
-      }
-    }
-    const i = r.next;
-    r.entries[i]?.marker?.dispose?.();
-    r.entries[i] = { marker, val };
-    r.next = (i + 1) % TS_POOL;
-  };
-  const tsEnsureLine = (term: XTerm, line: number, timestampsByLine: Map<number, string>) => {
-    if (term.buffer.active.type !== 'normal') return '';
-    const existing = timestampsByLine.get(line);
-    if (existing) return existing;
-    const currentLine = term.buffer.active.baseY + term.buffer.active.cursorY;
-    const ts = formatTerminalTimestamp();
-    const marker = term.registerMarker(line - currentLine);
-    tsSet(marker, ts);
-    if (marker) timestampsByLine.set(line, ts);
-    return marker ? ts : '';
-  };
-  const tsClearLine = (line: number) => {
-    const entries = tsRingRef.current.entries;
-    for (let i = 0; i < entries.length; i += 1) {
-      const entry = entries[i];
-      if (entry?.marker?.line === line) {
-        entry.marker.dispose?.();
-        entries[i] = null;
-      }
-    }
-  };
-  const tsClear = () => {
-    tsRingRef.current.entries.forEach((entry) => entry?.marker?.dispose?.());
-    tsRingRef.current.entries.fill(null);
-    tsRingRef.current.next = 0;
-  };
-  // 按 buffer 行号快照时间戳（收起/展开改写 buffer 后要还原，不能重新 now()）
-  // 与 syncGutter 一致：ring 从旧到新扫，后写覆盖，保留「执行时刻」而非提示符出现时刻
-  const tsSnapshotByLine = (term: XTerm, lineCount: number) => {
-    const total = typeof lineCount === 'number' ? lineCount : (term?.buffer?.active?.length || 0);
-    const byLine = new Array(Math.max(0, total)).fill('');
-    const ring = tsRingRef.current;
-    for (let offset = 0; offset < ring.entries.length; offset += 1) {
-      const index = (ring.next + offset) % ring.entries.length;
-      const entry = ring.entries[index];
-      const line = entry?.marker?.line;
-      if (!entry || entry.marker?.isDisposed || typeof line !== 'number' || line < 0 || line >= byLine.length) continue;
-      byLine[line] = entry.val;
-    }
-    return byLine;
-  };
-  const tsRemountFromList = (term: XTerm, tsList: string[]) => {
-    tsClear();
-    if (!term?.buffer?.active || !Array.isArray(tsList) || !timestampsEnabledRef.current) return;
-    const bufLen = term.buffer.active.length;
-    const cursorLine = term.buffer.active.baseY + term.buffer.active.cursorY;
-    const limit = Math.min(tsList.length, bufLen);
-    for (let line = 0; line < limit; line += 1) {
-      const val = tsList[line];
-      if (!val) continue;
-      try {
-        const marker = term.registerMarker(line - cursorLine);
-        if (marker) tsSet(marker, val);
-      } catch (_) {}
-    }
-  };
-  // 命令块：扫描 buffer 里「提示符行 → 下一提示符行」；收起时改写 buffer 只留一行摘要
-  // 同名命令用「第几次出现」区分，避免收起第二个把第一个状态冲掉
-  const CB_POOL = 400;
-  // null! 惰性初始化惯用法：下方 if 守卫保证首次渲染即完成填充，后续恒非空
-  const cbBlocksRef = useRef<Map<string, CommandBlockState>>(null!);
-  if (!cbBlocksRef.current) {
-    // key = `${commandLineText}#${occurrence}`；value = { id, commandLineText, occurrence, collapsed, savedOutput, savedOutputTs }
-    cbBlocksRef.current = new Map();
-  }
-  const cbIdSeqRef = useRef(1);
-  const cbRewriteLockRef = useRef(false);
-  const isCollapseSummaryLine = (text: string) => /^⋯\s+\d+\s+lines\s*$/.test(String(text || '').trim());
-  // 行首是 shell 提示符（后面可跟命令）。不能要求「整行以 # 结尾」，否则 `root@host:~# ping` 识别不到
-  const isShellPromptLine = (text: string) => {
-    const t = String(text || '').replace(/\s+$/g, '');
-    if (!t || t.length < 2 || isCollapseSummaryLine(t)) return false;
-    // user@host:path# cmd  /  user@host:path$ cmd
-    if (/^[\w.-]+@[\w.-]+:[^\n]*?[#\$](?:\s+|$)/.test(t)) return true;
-    // [user@host dir]$ cmd
-    if (/^\[[^\]]+\][#\$%](?:\s+|$)/.test(t)) return true;
-    // root@host ~]# cmd  一类
-    if (/^[\w.-]+@[\w.-]+\s+[^\n]*[#\$%](?:\s+|$)/.test(t)) return true;
-    // 极简：以 #/$ 单独起命令（少见）
-    if (/^[#\$]\s+\S/.test(t)) return true;
-    return false;
-  };
-  // 空提示符可以显示时间；真正执行（回车）时再更新该行时间戳
-  const normalizeCmdLineKey = (text: string) => String(text || '').replace(/\s+$/g, '');
-  const blockStateKey = (commandLineText: string, occurrence: number) => `${normalizeCmdLineKey(commandLineText)}#${occurrence}`;
-  const readTerminalBufferLines = (term: XTerm | null) => {
-    const buf = term?.buffer?.active;
-    if (!buf) return [];
-    const lines = [];
-    const total = buf.length;
-    for (let i = 0; i < total; i += 1) {
-      const bl = buf.getLine(i);
-      lines.push(bl ? bl.translateToString(true) : '');
-    }
-    while (lines.length > 1 && lines[lines.length - 1] === '') {
-      lines.pop();
-    }
-    return lines;
-  };
-  // 扫描所有提示符行下标：块 i = prompt[i] .. prompt[i+1]-1
-  const scanPromptIndexes = (lines: string[]) => {
-    const idxs: number[] = [];
-    for (let i = 0; i < lines.length; i += 1) {
-      if (isShellPromptLine(lines[i])) idxs.push(i);
-    }
-    return idxs;
-  };
-  // 在 buffer 里找「第 occurrence 次」出现的命令行（0-based）
-  const findCommandLineOccurrence = (lines: string[], commandLineText: string, occurrence: number) => {
-    const key = normalizeCmdLineKey(commandLineText);
-    let seen = 0;
-    for (let i = 0; i < lines.length; i += 1) {
-      if (normalizeCmdLineKey(lines[i]) !== key) continue;
-      if (seen === occurrence) return i;
-      seen += 1;
-    }
-    return -1;
-  };
-  const getOrCreateBlockState = (commandLineText: string, occurrence: number) => {
-    const textKey = normalizeCmdLineKey(commandLineText);
-    if (!textKey) return null;
-    const occ = Math.max(0, Number(occurrence) || 0);
-    const key = blockStateKey(textKey, occ);
-    const map = cbBlocksRef.current;
-    let block = map.get(key);
-    if (!block) {
-      if (map.size >= CB_POOL) {
-        const firstKey = map.keys().next().value;
-        if (firstKey != null) map.delete(firstKey);
-      }
-      block = {
-        id: cbIdSeqRef.current++,
-        commandLineText: textKey,
-        occurrence: occ,
-        collapsed: false,
-        savedOutput: null,
-        savedOutputTs: null,
-      };
-      map.set(key, block);
-    }
-    return block;
-  };
-  const rewriteTerminalBufferLines = (term: XTerm | null, lines: string[], nextTimestamps: string[] | null = null, options: { anchorLine?: number } = {}) => {
-    if (!term) return;
-    const anchorLine = typeof options.anchorLine === 'number' ? options.anchorLine : -1;
-    const bufBefore = term.buffer?.active;
-    const viewportBefore = bufBefore ? bufBefore.viewportY : 0;
-    const baseBefore = bufBefore ? bufBefore.baseY : 0;
-    // 尽量保持视口相对锚点行的偏移，减少重写后整屏「跳一下」
-    const anchorOffset = anchorLine >= 0 ? (anchorLine - viewportBefore) : 0;
-
-    cbRewriteLockRef.current = true;
-    // 一次 write 完成：清 scrollback + 清屏 + 回顶 + 写回。
-    // 比 term.reset() 轻（不重置模式/字符集），并用 write 回调在同一渲染路径里恢复滚动与 gutter。
-    const normalized = (Array.isArray(lines) ? lines : []).map((line) => String(line ?? ''));
-    const body = normalized.length === 0
-      ? ''
-      : normalized.length === 1
-        ? normalized[0]
-        : `${normalized.slice(0, -1).join('\r\n')}\r\n${normalized[normalized.length - 1]}`;
-    // \x1b[3J 清 scrollback，\x1b[2J 清屏，\x1b[H 光标回原点
-    const payload = `\x1b[3J\x1b[2J\x1b[H${body}`;
-
-    const finish = () => {
-      try {
-        if (Array.isArray(nextTimestamps)) {
-          tsRemountFromList(term, nextTimestamps);
-        }
-        const buf = term.buffer.active;
-        const maxVp = Math.max(0, buf.baseY);
-        if (anchorLine >= 0) {
-          const nextVp = Math.max(0, Math.min(maxVp, anchorLine - anchorOffset));
-          term.scrollToLine(nextVp);
-        } else if (baseBefore > 0) {
-          const ratio = Math.min(1, viewportBefore / baseBefore);
-          term.scrollToLine(Math.floor(maxVp * ratio));
-        } else {
-          term.scrollToBottom();
-        }
-      } catch (_) {}
-      cbRewriteLockRef.current = false;
-      // 直接 sync，少一帧 rAF 延迟，减轻「收起后 gutter 晚半拍」
-      try { syncGutter(); } catch (_) { scheduleGutterSync(); }
-    };
-
-    try {
-      // 只影响本地 xterm，不发给 SSH
-      term.write(payload, finish);
-    } catch (_) {
-      finish();
-    }
-  };
-  const cbToggleBlock = (blockId: number) => {
-    const term = termRef.current;
-    if (!term || term.buffer.active.type !== 'normal' || cbRewriteLockRef.current) return;
-    let block = null;
-    for (const b of cbBlocksRef.current.values()) {
-      if (b.id === blockId) { block = b; break; }
-    }
-    if (!block) return;
-
-    const lines = readTerminalBufferLines(term);
-    const oldTs = tsSnapshotByLine(term, lines.length);
-    // 用「命令文本 + 第几次出现」定位，避免两次 ping 绑到同一行
-    const start = findCommandLineOccurrence(lines, block.commandLineText, block.occurrence);
-    if (start < 0) return;
-
-    if (!block.collapsed) {
-      // 收起：start+1 到「下一提示符前」→ 换成一行摘要
-      const prompts = scanPromptIndexes(lines);
-      const pIdx = prompts.indexOf(start);
-      let end = start;
-      if (pIdx >= 0 && pIdx + 1 < prompts.length) {
-        end = prompts[pIdx + 1] - 1;
-      } else {
-        // 最后一个提示符：若下一行是摘要则用摘要；否则跟到末尾有内容的行
-        if (isCollapseSummaryLine(lines[start + 1])) {
-          end = start + 1;
-        } else {
-          end = start;
-          for (let i = start + 1; i < lines.length; i += 1) {
-            if (String(lines[i] || '').trim()) end = i;
-          }
-        }
-      }
-      if (end <= start) return;
-      // output / outputTs 同步构建，避免 filter 后 index 错位
-      const output = [];
-      const outputTs = [];
-      for (let i = start + 1; i <= end; i += 1) {
-        if (isCollapseSummaryLine(lines[i])) continue;
-        output.push(lines[i]);
-        outputTs.push(oldTs[i] || '');
-      }
-      if (output.length === 0) return;
-      block.savedOutput = output;
-      block.savedOutputTs = outputTs;
-      block.collapsed = true;
-      const nextLines = [
-        ...lines.slice(0, start + 1),
-        `⋯ ${output.length} lines`,
-        ...lines.slice(end + 1),
-      ];
-      // 严格 index 手术：前缀 + 摘要 + 后缀（后缀行号整体前移，戳也整段切开）
-      const summaryTs = outputTs.find(Boolean) || oldTs[start] || '';
-      const nextTs = [
-        ...oldTs.slice(0, start + 1),
-        summaryTs,
-        ...oldTs.slice(end + 1),
-      ];
-      // 以命令行作锚点，重写后视口尽量不跳
-      rewriteTerminalBufferLines(term, nextLines, nextTs, { anchorLine: start });
-      return;
-    }
-
-    // 展开
-    if (!Array.isArray(block.savedOutput) || block.savedOutput.length === 0) {
-      block.collapsed = false;
-      scheduleGutterSync();
-      return;
-    }
-    let summaryIdx = start + 1;
-    if (!isCollapseSummaryLine(lines[summaryIdx])) {
-      const nearby = lines.findIndex((l, idx) => idx > start && idx <= start + 3 && isCollapseSummaryLine(l));
-      if (nearby < 0) {
-        block.collapsed = false;
-        block.savedOutput = null;
-        block.savedOutputTs = null;
-        scheduleGutterSync();
-        return;
-      }
-      summaryIdx = nearby;
-    }
-    const restoredTs = Array.isArray(block.savedOutputTs) && block.savedOutputTs.length === (block.savedOutput || []).length
-      ? block.savedOutputTs
-      : (block.savedOutput || []).map(() => oldTs[start] || '');
-    const nextLines = [
-      ...lines.slice(0, start + 1),
-      ...block.savedOutput,
-      ...lines.slice(summaryIdx + 1),
-    ];
-    const nextTs = [
-      ...oldTs.slice(0, start + 1),
-      ...restoredTs,
-      ...oldTs.slice(summaryIdx + 1),
-    ];
-    block.collapsed = false;
-    rewriteTerminalBufferLines(term, nextLines, nextTs, { anchorLine: start });
-  };
-  // 关闭功能前：把所有已收起的块展开回 buffer，否则 savedOutput 清掉后无法再展开
-  const cbExpandAllCollapsed = (term: XTerm | null) => {
-    if (!term || term.buffer.active.type !== 'normal' || cbRewriteLockRef.current) return false;
-    const collapsed = [...cbBlocksRef.current.values()].filter(
-      (b) => b && b.collapsed && Array.isArray(b.savedOutput) && b.savedOutput.length > 0,
-    );
-    if (collapsed.length === 0) return false;
-
-    let lines = readTerminalBufferLines(term);
-    let oldTs = tsSnapshotByLine(term, lines.length);
-    // 从后往前展开，避免前面插入行导致后面 occurrence 定位错位
-    collapsed.sort((a, b) => b.occurrence - a.occurrence || b.commandLineText.localeCompare(a.commandLineText));
-
-    for (const block of collapsed) {
-      const start = findCommandLineOccurrence(lines, block.commandLineText, block.occurrence);
-      if (start < 0) {
-        block.collapsed = false;
-        block.savedOutput = null;
-        block.savedOutputTs = null;
-        continue;
-      }
-      let summaryIdx = start + 1;
-      if (!isCollapseSummaryLine(lines[summaryIdx])) {
-        const nearby = lines.findIndex((l, idx) => idx > start && idx <= start + 3 && isCollapseSummaryLine(l));
-        if (nearby < 0) {
-          block.collapsed = false;
-          block.savedOutput = null;
-          block.savedOutputTs = null;
-          continue;
-        }
-        summaryIdx = nearby;
-      }
-      const restoredTs = Array.isArray(block.savedOutputTs) && block.savedOutputTs.length === (block.savedOutput || []).length
-        ? block.savedOutputTs
-        : (block.savedOutput || []).map(() => oldTs[start] || '');
-      lines = [
-        ...lines.slice(0, start + 1),
-        ...(block.savedOutput || []),
-        ...lines.slice(summaryIdx + 1),
-      ];
-      oldTs = [
-        ...oldTs.slice(0, start + 1),
-        ...restoredTs,
-        ...oldTs.slice(summaryIdx + 1),
-      ];
-      block.collapsed = false;
-      block.savedOutput = null;
-      block.savedOutputTs = null;
-    }
-    rewriteTerminalBufferLines(term, lines, oldTs);
-    return true;
-  };
-  const cbClear = () => {
-    cbBlocksRef.current = new Map();
-  };
-  const gutterRef = useRef<HTMLDivElement | null>(null);
-  const gutterSyncRAFRef = useRef<number | null>(null);
-  const linkUnderlineLayerRef = useRef<HTMLDivElement | null>(null);
-  const linkUnderlineSyncRAFRef = useRef<number | null>(null);
+  // ── 时间戳 ring / 链接下划线 / 命令块 gutter（terminal/ 目录同名 hook） ──
+  const { tsRingRef, tsSet, tsEnsureLine, tsClearLine, tsClear, tsSnapshotByLine, tsRemountFromList } = useTerminalTimestamps({ timestampsEnabledRef });
+  const { linkUnderlineLayerRef, linkUnderlineSyncRAFRef, getViewportLinkCache, scheduleLinkUnderlineSync } = useTerminalLinkUnderlines({ termRef, containerRef });
+  const { gutterRef, gutterSyncRAFRef, cbRewriteLockRef, isShellPromptLine, isCollapseSummaryLine, scheduleGutterSync, syncGutter, handleClearScreen, cbExpandAllCollapsed, cbClear } = useTerminalGutter({
+    termRef,
+    containerRef,
+    timestampsEnabledRef,
+    commandBlocksEnabledRef,
+    alternateBufferActiveRef,
+    commandBlocksVisible,
+    t,
+    tsRingRef,
+    tsEnsureLine,
+    tsClearLine,
+    tsSnapshotByLine,
+    tsRemountFromList,
+  });
   const smartWriteRef = useRef<((data: string | Uint8Array) => void) | null>(null);
 
   // ponytail: getTerminalTheme() 每次渲染调用 30+ 次，缓存为 1 次
   const T = useMemo(() => getTerminalTheme(), [themeToggle]);
 
-  // ── 时间轴 / 命令块：同步 gutter 到 xterm 视口 ─────────────────
-  function scheduleGutterSync() {
-    const gutterNeeded = timestampsEnabledRef.current || commandBlocksEnabledRef.current;
-    if (gutterSyncRAFRef.current !== null || !gutterNeeded || alternateBufferActiveRef.current) return;
-    gutterSyncRAFRef.current = requestAnimationFrame(() => {
-      gutterSyncRAFRef.current = null;
-      syncGutter();
-    });
-  }
 
-  // ── 链接：可见区扫描缓存（下划线与 provider 共用，避免双扫） ────
-  const viewportLinkCacheRef = useRef({ key: '', byLine: new Map<number, Array<{ text: string; range: IBufferRange }>>() });
-
-  function getViewportLinkCache(term: XTerm) {
-    const buf = term.buffer.active;
-    const rows = term.rows || 0;
-    const viewportY = buf.viewportY;
-    // 简单 key：视口位置 + 行数 + 输入行起点（输入行变化时失效）
-    const inputStart = getTerminalInputStartLine(term);
-    const key = `${viewportY}|${rows}|${inputStart}|${buf.baseY}|${buf.cursorY}`;
-    const cache = viewportLinkCacheRef.current;
-    if (cache.key === key) return cache.byLine;
-    const byLine: Map<number, Array<{ text: string; range: IBufferRange }>> = new Map();
-    for (let row = 0; row < rows; row += 1) {
-      const bufferLineNumber = viewportY + row + 1;
-      const links = findTerminalHttpLinksOnLine(term, bufferLineNumber);
-      if (links.length) byLine.set(bufferLineNumber, links);
-    }
-    viewportLinkCacheRef.current = { key, byLine };
-    return byLine;
-  }
-
-  function invalidateViewportLinkCache() {
-    viewportLinkCacheRef.current = { key: '', byLine: new Map() };
-  }
-
-  function scheduleLinkUnderlineSync() {
-    if (linkUnderlineSyncRAFRef.current !== null) return;
-    linkUnderlineSyncRAFRef.current = requestAnimationFrame(() => {
-      linkUnderlineSyncRAFRef.current = null;
-      invalidateViewportLinkCache();
-      syncLinkUnderlines();
-    });
-  }
-
-  function syncLinkUnderlines() {
-    const layer = linkUnderlineLayerRef.current;
-    const term = termRef.current;
-    const container = containerRef.current;
-    if (!layer) return;
-    if (!term?.buffer?.active || !container) {
-      layer.innerHTML = '';
-      return;
-    }
-    const screen = container.querySelector('.xterm-screen');
-    const rowsEl = container.querySelector('.xterm-rows');
-    if (!screen || !rowsEl) {
-      layer.innerHTML = '';
-      return;
-    }
-    const cols = term.cols || 1;
-    const rows = term.rows || 1;
-    const screenRect = screen.getBoundingClientRect();
-    const cellWidth = screenRect.width / cols;
-    const cellHeight = screenRect.height / rows;
-    const rowsRect = rowsEl.getBoundingClientRect();
-    const offsetX = rowsRect.left - screenRect.left;
-    const offsetY = rowsRect.top - screenRect.top;
-    const byLine = getViewportLinkCache(term);
-    const parts: string[] = [];
-    const seen = new Set();
-    const viewportY = term.buffer.active.viewportY;
-    byLine.forEach((links) => {
-      for (const link of links) {
-        const id = `${link.range.start.y}:${link.range.start.x}-${link.range.end.y}:${link.range.end.x}:${link.text}`;
-        if (seen.has(id)) continue;
-        seen.add(id);
-        const y0 = link.range.start.y;
-        const y1 = link.range.end.y;
-        for (let by = y0; by <= y1; by += 1) {
-          const row = by - 1 - viewportY;
-          if (row < 0 || row >= rows) continue;
-          const startCol = by === y0 ? Math.max(0, link.range.start.x - 1) : 0;
-          const endCol = by === y1 ? Math.max(startCol + 1, link.range.end.x) : cols;
-          const left = offsetX + startCol * cellWidth;
-          const width = Math.max(cellWidth, (endCol - startCol) * cellWidth);
-          const top = offsetY + (row + 1) * cellHeight - 2;
-          parts.push(
-            `<div style="position:absolute;left:${left}px;top:${top}px;width:${width}px;height:0;border-bottom:1px solid var(--accent, #4d9eff);pointer-events:none;"></div>`,
-          );
-        }
-      }
-    });
-    layer.innerHTML = parts.join('');
-  }
-
-  function collectLiveCommandBlocks(term: XTerm) {
-    // 完全按 buffer 扫描提示符：prompt[i] .. prompt[i+1]-1
-    // 同名命令按出现次序分块，避免两次 ping 共用状态
-    const lines = readTerminalBufferLines(term);
-    const prompts = scanPromptIndexes(lines);
-    const list = [];
-    const occurrenceByText = new Map();
-    for (let i = 0; i < prompts.length; i += 1) {
-      const start = prompts[i];
-      const commandLineText = normalizeCmdLineKey(lines[start]);
-      const occurrence = occurrenceByText.get(commandLineText) || 0;
-      occurrenceByText.set(commandLineText, occurrence + 1);
-      const block = getOrCreateBlockState(commandLineText, occurrence);
-      if (!block) continue;
-      let end = start;
-      if (block.collapsed && isCollapseSummaryLine(lines[start + 1])) {
-        end = start + 1;
-      } else if (i + 1 < prompts.length) {
-        end = Math.max(start, prompts[i + 1] - 1);
-      } else if (isCollapseSummaryLine(lines[start + 1])) {
-        end = start + 1;
-      } else {
-        end = start;
-        for (let j = start + 1; j < lines.length; j += 1) {
-          if (isShellPromptLine(lines[j])) break;
-          if (String(lines[j] || '').trim()) end = j;
-        }
-      }
-      list.push({ block, start, end, collapsed: Boolean(block.collapsed) });
-    }
-    return list;
-  }
-
-  function syncGutter() {
-    const gutter = gutterRef.current;
-    const term = termRef.current;
-    const showTs = timestampsEnabledRef.current;
-    const showCb = commandBlocksEnabledRef.current;
-    if (!gutter || !term || (!showTs && !showCb) || term.buffer.active.type !== 'normal') {
-      return;
-    }
-    const buf = term.buffer.active;
-    const rows = term.rows;
-    if (!rows || !containerRef.current) return;
-
-    const timestampsByLine = new Map();
-    if (showTs) {
-      const ring = tsRingRef.current;
-      // 从旧到新扫：后写覆盖先写，保证「执行时刻」压过「提示符出现时刻」
-      for (let offset = 0; offset < ring.entries.length; offset += 1) {
-        const index = (ring.next + offset) % ring.entries.length;
-        const entry = ring.entries[index];
-      const line = entry?.marker?.line;
-      if (!entry || entry.marker.isDisposed || typeof line !== 'number' || line < 0) {
-        ring.entries[index] = null;
-      } else {
-        timestampsByLine.set(line, entry.val);
-      }
-      }
-    }
-
-    const liveBlocks = showCb ? collectLiveCommandBlocks(term) : [];
-    // 行 → 所在块；块起点优先
-    const blockByLine = new Map();
-    liveBlocks.forEach((item) => {
-      for (let line = item.start; line <= item.end; line += 1) {
-        const existing = blockByLine.get(line);
-        if (existing && existing.start === line && existing.start !== item.start) continue;
-        blockByLine.set(line, item);
-      }
-    });
-
-    const firstVisible = buf.viewportY; // buffer 中第一个可见行 (ydisp)
-
-    // 通过 xterm screen/rows 的实际渲染尺寸计算行高，确保像素级对齐
-    const screen = containerRef.current.querySelector('.xterm-screen');
-    const rowsEl = containerRef.current.querySelector('.xterm-rows');
-    let lineH = 0;
-    if (screen && rowsEl) {
-      const screenRect = screen.getBoundingClientRect();
-      const rowsRect = rowsEl.getBoundingClientRect();
-      lineH = Math.max(rowsRect.height / rows, 1);
-      const paddingTop = `${Math.max(rowsRect.top - screenRect.top, 0)}px`;
-      if (gutter.style.paddingTop !== paddingTop) gutter.style.paddingTop = paddingTop;
-    } else {
-      // xterm options 类型为可选，运行期恒有默认值（13 / 1.22）
-      lineH = (term.options.fontSize ?? 13) * (term.options.lineHeight ?? 1.2);
-    }
-
-    // 时间戳用状态色；命令块树线/折叠钮用 accent，深色终端上更醒目
-    const tsColor = 'var(--term-status-color)';
-    const blockColor = 'var(--accent)';
-    let html = '';
-    for (let i = 0; i < rows; i++) {
-      const tsIdx = firstVisible + i;
-      const bufLine = buf.getLine(tsIdx);
-      const lineText = bufLine ? bufLine.translateToString(true) : '';
-      const isEmptyLine = !bufLine || lineText === '';
-      const isWrapped = bufLine && bufLine.isWrapped;
-      let ts = '';
-      // 空提示符也可显示已有戳；当前光标行没有戳时补一个（出现提示符的时间）
-      // 真正执行命令时会在回车路径更新为执行时刻
-      if (showTs && !isEmptyLine && !isWrapped && tsIdx >= 0 && !isCollapseSummaryLine(lineText)) {
-        ts = timestampsByLine.get(tsIdx)
-          || (tsIdx === buf.baseY + buf.cursorY ? tsEnsureLine(term, tsIdx, timestampsByLine) : '');
-      }
-
-      const owning = showCb ? blockByLine.get(tsIdx) : null;
-      const parts = [];
-      // 固定列宽：时间戳列右对齐贴齐「]」，命令块列固定 14px，中间无 gap，避免看起来多一格空
-      if (showTs) {
-        // [HH:MM:SS] 共 10 字符；等宽 11px 约 66px，固定 70 够用
-        parts.push(`<span style="display:inline-block;width:70px;min-width:70px;max-width:70px;text-align:right;flex-shrink:0;letter-spacing:0;box-sizing:border-box;color:${tsColor}">${ts || ''}</span>`);
-      }
-      if (showCb) {
-        // 纯 CSS 树线：连续竖条 + 末行 L 折角（不用 │/└ 字符，避免粗细/对齐怪）
-        // 整数 2px + border 画 L，避免 1.5px 横线抗锯齿后显得更细
-        const barBg = `color-mix(in srgb, ${blockColor} 92%, transparent)`;
-        const barW = 2;
-        const armW = 7;
-        const cell = (inner: string) =>
-          `<span style="position:relative;display:inline-block;width:14px;min-width:14px;height:${lineH}px;flex-shrink:0;box-sizing:border-box;vertical-align:top">${inner || ''}</span>`;
-        // 竖条水平居中于 14px 列
-        const vBar = (top: string, bottom: string) =>
-          `<span style="position:absolute;left:50%;top:${top};bottom:${bottom};width:${barW}px;margin-left:-${barW / 2}px;background:${barBg};pointer-events:none"></span>`;
-        // 末行 L：同宽 border-left + border-bottom；横臂贴文字垂直中心略偏下（约 0.55em 基线感）
-        // 用 lineH 比例 + 字号无关的下限，大字号时不显得折角悬在行上半
-        const cornerH = Math.max(barW + 4, Math.round(lineH * 0.72));
-        const lCorner = () =>
-          `<span style="position:absolute;left:50%;top:0;width:${armW}px;height:${cornerH}px;margin-left:-${barW / 2}px;border-left:${barW}px solid ${barBg};border-bottom:${barW}px solid ${barBg};box-sizing:border-box;pointer-events:none"></span>`;
-        let blockCell = cell('');
-        if (owning && tsIdx === owning.start) {
-          // 可折叠：展开有输出，或已收起可再展开
-          const canFold = owning.collapsed || owning.end > owning.start;
-          const icon = owning.collapsed ? '+' : '−';
-          if (canFold) {
-            const padTop = Math.max(0, (lineH - 14) / 2);
-            const btn = `<button type="button" data-cb-id="${owning.block.id}" title="${owning.collapsed ? t('展开') : t('收起')}" style="position:absolute;left:0;top:${padTop}px;display:inline-flex;align-items:center;justify-content:center;width:14px;min-width:14px;height:14px;margin:0;padding:0;border:1px solid color-mix(in srgb, ${blockColor} 78%, transparent);border-radius:2px;background:color-mix(in srgb, ${blockColor} 16%, transparent);color:${blockColor};font-size:11px;line-height:1;cursor:pointer;font-family:var(--font-mono);box-sizing:border-box;font-weight:700;z-index:1">${icon}</button>`;
-            // 多行展开：按钮下接到行底，与下一行整高竖条无缝衔接
-            const barBelow = (!owning.collapsed && owning.end > owning.start)
-              ? vBar(`${padTop + 14}px`, '0')
-              : '';
-            blockCell = cell(`${btn}${barBelow}`);
-          } else {
-            const padTop = Math.max(0, (lineH - 14) / 2);
-            blockCell = cell(`<span style="position:absolute;left:0;top:${padTop}px;display:inline-flex;width:14px;height:14px;align-items:center;justify-content:center;border:1px solid color-mix(in srgb, ${blockColor} 55%, transparent);border-radius:2px;opacity:0.75;box-sizing:border-box"></span>`);
-          }
-        } else if (owning && !owning.collapsed && tsIdx > owning.start && tsIdx < owning.end) {
-          blockCell = cell(vBar('0', '0'));
-        } else if (owning && !owning.collapsed && tsIdx === owning.end && owning.end > owning.start) {
-          blockCell = cell(lCorner());
-        }
-        parts.push(blockCell);
-      }
-      // overflow:visible：命令块竖条才能顶满行高无裁切；时间戳等宽不溢出
-      html += `<div style="height:${lineH}px;line-height:${lineH}px;font-size:11px;font-family:var(--font-mono);font-variant-numeric:tabular-nums;white-space:nowrap;overflow:visible;padding:0 2px 0 4px;box-sizing:border-box;display:flex;align-items:center;justify-content:flex-end;gap:2px">${parts.join('')}</div>`;
-    }
-    gutter.innerHTML = html;
-  }
-
-  // gutter 点击折叠/展开（随显示状态重绑，避免首屏 display:none 时漏挂）
-  useEffect(() => {
-    const gutter = gutterRef.current;
-    if (!gutter || !commandBlocksVisible) return undefined;
-    const onClick = (event: MouseEvent) => {
-      const btn = (event.target as Element | null)?.closest?.('button[data-cb-id]');
-      if (!btn) return;
-      event.preventDefault();
-      event.stopPropagation();
-      const id = Number(btn.getAttribute('data-cb-id'));
-      if (Number.isFinite(id)) cbToggleBlock(id);
-    };
-    gutter.addEventListener('click', onClick);
-    return () => gutter.removeEventListener('click', onClick);
-  }, [commandBlocksVisible]);
-
-  // ── 终端清屏处理：清空视口对应的时间戳 ─────────────────────
-  function handleClearScreen() {
-    const term = termRef.current;
-    if (!term || term.buffer.active.type !== 'normal') return;
-    const buf = term.buffer.active;
-    const rows = term.rows || 24;
-    const firstVisible = buf.viewportY;
-    for (let i = 0; i < rows; i++) {
-      tsClearLine(firstVisible + i);
-    }
-    scheduleGutterSync();
-  }
 
   // ── 初始化 xterm + WebSocket 终端通道 ────────────────────────────────
   // xterm.js 通过 AttachAddon + WebSocket 直接连到本地 Go WebSocket 服务器
